Fix FactoryGirl errors when running migrations
Prior to this with the gem set to fale we lost several useful features of actually having the FactoryGirl gem installed, especially around generators.

Bringing the gem back the following error returned:

    rake aborted!
    Sequel::DatabaseConnectionError: Mysql2::Error: Unknown database 'saml_development'
    ...

This commit ensures that factories which use the :class symbol put the names of the associated classes in quotes. This corrects the migration problem above in both development and test. Future factories using the :class symbol must also put the class name in quotes.