Refine modeling of CA certificates and info
Due to complications with establishing a low level database relationship between the previously modelled CAKeyInfo and CACertificate this implementation refines the implementation to store all certificate data within the CAKeyInfo model.

This is appropriate as any real world case we've ever worked with directly maps an xmldig KeyInfo element to a single X509Data element.

As the xmldsig schema strictly does allow for multiple X509Data blocks this model will need to be extended should this ever be required. Should the the non modelled elements RetrievalMethod, PGPData, SPKIData or MgmtData be required this model will also need to be extended.