Model XMLDSig KeyInfo/Certificate
Due to complications with establishing a low level database relationship between KeyInfo and Certificate this implementation stores all certificate data within the KeyInfo model.

This is appropriate as any real world case we've ever worked with directly maps an xmldig KeyInfo element to a single X509Data element.

As the xmldsig schema strictly does allow for multiple X509Data blocks this model will need to be extended should this ever be required. Should the the non modelled elements RetrievalMethod, PGPData, SPKIData or MgmtData be required this model will also need to be extended.