Allow possible Roots to access publication_info
Within eduGAIN metadata producer requirements specification the
need to render a PublicationInfo node exists for all metadata root
elements.

As SAML Metadata 2.0 specifies that both EntitiesDescriptor and
EntityDescriptor may play the role of root node both of these need
access to PublicationInfo.

Given each deployment of saml-service will have an
EntitiesDescriptor instance that acts as the top of the tree and is
also likely to be the only object actually storing a PublicationInfo
reference we need to be able to access this data should we choose to
render metadata from some lower point in the tree, in particular a
single EntityDescriptor would be the most common use case here.

The added method #locate_publication_info in partnership with existing
validations supports this.