Implementation of MDQP as first public interface
This commit implements:
https://tools.ietf.org/pdf/draft-young-md-query-05.pdf and a portion
of https://tools.ietf.org/pdf/draft-young-md-query-saml-05.pdf due to
overlap in current use-case. The remainder of the SAML profile,
largely 2.2.2, Transformed Identifiers will be supplied in a
future commit.

MDQP defines a simple protocol for retrieving metadata about
named entities, or named collections of entities. The goal of the
protocol is to profile various aspects of HTTP to allow requesters to
rely on certain, rigorously defined, behaviour.

This represents the first method of public access to metadata content
that saml-service will support. Given the flexibility it affords it
may be the only way we publicly expose metadata content.

We expose entities as follows:

* All entities associated with a metadata instance if that instance
  has agreed to publish a total aggregate
* All entities associated with a metadata instance that are also
  identified by some known tag e.g. All IdP by querying for the 'idp'
  identifier. Tags are simple words that have meaning defined by the
  owner of each metadata instance.
* A specific entity identified by the URI representing a SAML entityID