Remove duplicate entity_id String from KnownEntity
This data is stored within the relevant EntityDescriptor or
RawEntityDescriptor EntityID relationship which is maintained by
KnownEntity.

As such it no longer makes sense to duplicate this data within the
known_entities table. Previously RawEntityDescriptor had no EntityId
relationship which is why this data was originally stored with
KnownEntity.

For ease of use KnownEntity now implements #entity_id which mimics the
previous behaviour of providing the entity_id uri which it ultimately
represents.

@smangelsdorf please cast an extra critical eye over this commit
during MDQP review to ensure something important was not missed. Note
especially the removal of 2 assertions in update_entity_source_spec
that no longer seemed to make sense.