Tag new FR sourced EntityDescriptors with fed tag
In order to correctly utilise MDQP endpoints all federation ED should
be tagged with a value represented within the associated
MetadataInstance as 'primary_tag'.

The value configured for jobs importing entities will need to match
that of the MetadataInstance in use by the federation. In the case of
the AAF this is expected to be 'aaf'.