Tag new remotely sourced ED with fed/source tag
In order to correctly utilise MDQP endpoints all ED KE should
be tagged to indicate their primary source federation.

In the case of the AAF usage of eduGAIN sources for example this is
expected to be 'edugain'. There may or may not be a MetadataInstance
whose primary key is also 'edugain' but this is likely to be how the
AAF will operate.