Ensure KE updated_at changes when jobs load data
As KnownEntity#updated_at is a key piece of the ETag generation logic
for MDQP it is critical that this value is updated anytime a piece of
data within its EntityDescriptor or RawEntityDescriptor is changed.

This commit adds the sequel touch plugin to support this function and
ensures that both FR and EntitySource jobs invoke touch for
each KnownEntity they interact with.