Refine contact validations on EntityDescriptor
Previously our implementation validated that a technical contact
existed to be compliant to edugain needs. However this is over
and above what the SAML specification mandates and isn't going to be
fulfilled by all existing sources, especially AAF FR which never
enforced this need.

Thus the hard requirement on edugain compatible instances has been
relaxed back to what SAML metadata core requires,
`<ContactPerson> [Zero or More]`.

To ensure we keep the edugain compliance validations however
additional logic in the form of `edugain_compliant?` has been added to
EntityDescriptor and all RoleDescriptors. This will be a useful future
check when considering tagging an EntityDescriptor for inclusion in
feeds destined for the edugain aggregator.