Prevent erroneous FR keys from tearing down import
Integration testing showed a single key stored within FR that was
somehow not ASN compliant even though those kinds of errors should be
prevented on ingest. It is possible that Java crypto on the FR side is
more forgiving than OpenSSL when parsing these certs.

To prevent a failed certificate from destroying an otherwise
healthly sync process we now capture OpenSSL exceptions, report them
via the Rails logger and carry on.

Additionally logging for the Entity being ingested was added to the
start of the process as this was helpful when debugging sync failure.