Allow MDQ more time to complete requests
When a large number of entities are present (i.e. entire eduGAIN set
of ~2775 nodes) requests can take longer than 60 seconds to complete
if not cached.

This will stop unicorn killing the request as 'dead'.

Longer term we should bring this up with MDQ folk, perhaps there is
some response we could send if a cached version is not ready and the
responder considers it will take to long to complete building for a
standard HTTP timeout.