Add `source_tag` field to EntitySource
This tag value MUST be automatically applied to all entities
originating from this source. The responsibility for this is left to
the implementation of UpdateFromX jobs (see next commit).

Additionally this commit marks the `source_tag` field as both unique
and required. The default value of `aaf` will allow the migration to
proceed in both the test and production federations which have already
deployed the current schema but only have a single EntitySource
instance.