Ensure ED from lowest ranked EntitySource used
Per review notes there was a spec missing to validate that ED from the
lowest ranked entity source was indeed used.

Specs now ensure that when duplicate ED are held and tagged with the
same values that the version from the lowest ranked
(i.e. most trusted) EntitySource is utilised in rendering metadata.