Ensure SAML metadata ns is always xmlns default
Previously remote sources of metadata had their own ideas about
scoping for the uri 'urn:oasis:names:tc:SAML:2.0:metadata'.

Internally we expect this to be the default prefix for xmlns but other
sources, especially eduGAIN scope this to the prefix 'xmlns:md'.

On ingest we now correct source XML so it adheres to our world view,
which allows us to publish seemlessly a mix of internally and
externally sourced entities.