Refine approach for DB pools before/after fork
After looking into the nature of #80 this is an initial attempt at
resolving the error state by ensuring that pools for workers are
robustly established when unicorn forks.

This is not necessarily considered to be an absolute fix but will
provide further insight to the fault in terms of diagnosis.