Ensure MDUI descriptions are valid
In production we discovered an issue where data from FR
for MDUI descriptions contained the encoded value &#13;
due, we believe, to folks pasting into the multi-line
description field from word docs.

This in itself did not pose a problem but the Shibboleth SP
does not correctly handle this case when writing out backup
files thus signature validation would fail when the SP tried
to restart when upstream was not modified.