Force usage of our patched Nokogiri
We require the patches supplied by the PR:

* https://github.com/sparklemotion/nokogiri/pull/1333 OR
* https://github.com/sparklemotion/nokogiri/pull/1444

In order to close #106 and continue to roll out eduGain support.

Unfortunately efforts to get these fixes merged have stalled
for a considerable amount of time now and we need to progress.

This change should be able to be reverted to stock Nokogiri once
1.6.8 is released, assuming 1333 or 1444 has been merged.

Closes #106.