Persist RawEntityDescriptor and associated models
Validates parameters and persists:
- A new RawEntityDescriptor
- A new KnownEntity
- A new EntityId

The "enabled" field is used from the payload when creating these models.

Updated/created timestamps in the request are no longer used. Therefore
the created/updated timestamps will be as per default. This means we do
not store when an IdP was created (happens in Hosted IdP Service). From
some exploration this shouldn't be an issue.

Right now when a RawEntityDescriptor is created the response code is
201. At this point a location header cannot be included (i.e. GET
/api/entity_sources/:tag/raw_entity_descriptors/:id as this resource
doesn't exist yet — something that may be required in the future.