Add entity_source_id for entity_ids unique key
We need to constrain each EntityId to ensure it is unique per
EntitySource. To that end, we need to duplicate the foreign key
`entity_source_id` in the `entity_ids` table.

The unique key applies to the SHA1 of the entityID URI, rather than to
the URI itself. This is necessitated by the multi-table inheritance
structure of the database, which ultimately gives us 3 options:

* Add a unique key on `entity_source_id` + `sha1` which are both in the
  `entity_ids` table
* Add `entity_source_id` to the `saml_uris` table instead, and have the
  unique key on `entity_source_id` + `uri`. This would require that
  *ALL* SamlUri objects are unique per EntitySource, which is not a
  constraint that we can reasonably impose
* Add the `entity_id` to the `known_entities` table. This was the
  original state of the models, but was reversed in 7e31466 and
  discussed in the context of #34. Although I don't recall the exact
  motivation here, it's also in direct contradiction to the wording of
  #122 which seems to imply this is less than ideal.

This can be refactored at such a time as the database is restructured to
a more simplified model.

Note: This also makes the uniqueness case-sensitive, which it would be
in either case given our `utf8-bin` collation settings.