Refactor to correctly maintain tags/flags on RED
When a RawEntityDescriptor was updated using `set_saml_type`, the tags
and flags weren't fully maintained as expected. In particular, if an
entity source unpublished an entity, and then published a new entity
with the same entityID but a different type, the data in saml-service
may have been adversely impacted.

We now ensure the correct set of tags and flags whenever `set_saml_type`
is called. Some related refactoring was done to tidy up the class, and a
full set of tests specific to the concern added to ease the refactoring
process.