Refactor etag helpers to avoid collisions
When using the set of known entities to cache a metadata document, it
creates a race condition when two metadata instances have the same set
of entities but a different signing key. The first document to be cached
will be served for both metadata instances, in that case.

To resolve that issue, the metadata instance identifier needs to be an
input to the etag generation.

Additionally, this combines both etag helpers into a single method.