Correct fault when json key not presented
Under production workloads FR treats `specificationRequired` as being
false if not provided and thus does not offer it in most JSON exports.

The code developed in the previous PR works in the opposite direction
expecting that the JSON schema will always be fully populated.

This commit introduces a default value when `specificationRequired` is
not presented to address the `ImplicitSchema::ValidationError`
which was being reported.