Ensure metadata queries are ordered by database id
This commit ensures that all collections are explicitly ordered
by the associated Rails ID when XML is being created.

In addition to being a failsafe against some future database change
this behaviour is how we approached ordering within Federation Registry
[https://github.com/ausaccessfed/federationregistry/blob/develop/app/plugins/metadata/grails-app/services/aaf/fr/metadata/MetadataGenerationService.groovy]
which seems like a reasonable default to take forward in the absence of
any better option.