Rework boolean for specification to be explicit
Previously we relied on `nil.blank?` to return false (fine) but also
an actual value of false in the api payload to have `false.blank?`
return true which is suboptimal if `blank?` ever changes somehow.

Now we default to `false` instead of `nil` if non existant and
specifically look for a match against false.