Correctly handle boolean parameters via API
Real clients send the request body as JSON, but in a Rails controller
test the payload is delivered (or simulated as being delivered) as a
form encoded request body. This has the effect that all values are
converted to strings, instead of being serialized / deserialized
correctly like a JSON payload would be.

Passing the `as: :json` argument to the test request method (`patch` in
this case) restores the correct behaviour and delivers real boolean
values to the application.