Correctly handle known_entities with no data
For a long-running saml-service, it's possible for KnownEntity records
to exist with no data underneath. They have no impact on the operation
of saml-service, but don't work correctly with the Git repository
export. This change causes the blank KnownEntity records to be skipped
just as they are during metadata generation.