Remove KeyName from AAF metadata
Having reviewed AAF metadata and the associated schema specifications it
was determined that KeyName was adding no value and could, under the
right conditions, be problematic.

This change will need to be previewed in the AAF test federation for a
reasonable amount of time to ensure that no breakage occurs for obscure
implementations, however this is not anticipated.