Support longer KeyInfo Subject and Issuer values
During production operation of SAML Service we identified a data
migration failure causing total transaction rollback when importing from
FR.

Investigation showed that CSIRO had provided a KeyInfo Subject line that
far exceeded our previous 255 limit.

As this is essentially a game of 'whack-a-mole' with regards to what is
the correct column width to support here I have gone for the win and
taken this right out to 2048, which at the length of a short novel and
far greater than that allowed in even a modern tweet, should see us
avoid this issue in the future.