Support multiple scopes for IDPSSODescriptor
An AAF customer had a legitimate need to support multiple scopes within
their metadata snippet. This was originally achieved with a regexp scope
but the approach was rejected by international federations that filter
metadata including JISC and Incommon.

This commit accepts a single value from FR for scope (which is all it
can provide) and permits additional 'locked' scopes to be manually
entered into the SAML service database if necessary to provide second,
third etc scopes in metadata.