WIP: Require explicit conf for API auth method
Need to ensure that secure posture for differing deployment environments
is able to be explicitly stated by OPS teams instead of guessed at based
on headers supplied in requests.

The previous approach was further weakend by the need to have upstream
load balancers blacklist certain headers from the wider internet (fine
in of itself but not ideal in this scenario).

WIP: Broken specs currently.