Fix: MDQCtrl: respect rank in specific_entity methods
So far, the specific_entity and specific_entity_sha1 methods in
MetadataQueryController were selecting the `.first` EntityId object found.

However, when an entity is defined in multiple entity sources,
this would pick the first one defined - which would break when
an entity is first defined in a lower-priority entity source (eduGAIN)
and only later in a high-priority entity source (FR).

So instead explicitly select the EntityId object connected
to the lowest rank (highest priority) entity source.