Fix flake in RawEntityDescriptorsController spec (#252)
The test on line 410 often failed because one or two of the new
tags sent by the controller might already be on the record. The
test then expects the updated record to be tagged twice.

This commit fixes the issue by manually specifying the pre-existing
tags and the new tags sent in the update request. This removes the
possibility of collision.