new: Spec: FR-sync/RequestedAttr: test only approved attributes are imported
Adjust tests to also test that only approved attributes are imported.

Add approved flag to the test `attribute_instances` data and use that for `requested_attribute_json`.

Add an extra element to attribute_instance with `approved: false`.

In test `requests expected number of attributes`, the expected value stays
the same (attribute_count + 1), expecting the unapproved attribute not to count.

However, the sanity check in `is provided with attributes that are not acceptable`
needs to adjust with the extra unapproved attribute to `attribute_count + 3`