fix: FR-sync: filter out requestedAttrs requiring spec with no requested values earlier
Fix the situation where an SPSSODescriptor with a a solo eduPersonEntitlement
requestedAttribute with no value breaks metadata generation - the requestedAttribute is
filtered out (for being an attribute that requires specification but not providing a value),
but this happens too late and an AttributeConsumingService is already created.

As SAML Metadata specification does not allow an AttributeConsumingService with no RequestedAttribute,
the rendered metadata fails XML schema validation.

Fix this by bringing the `correctly_specified?` check forward, filtering the attributes
before the AttributeConsumingService is created.

Add a test covering this situation to make sure that an SSPSSODescriptor
"with a solo requestedAttribute requiring specification with no requested value provided"
does not create a new AttributeConsumingService.