Make EntitySource signature verification optional (#332)
Can be redundant in the age of HTTPS.

Having a signature within the document is important if the document gets
downloaded to one place and passed around from there to other consumers.
But if you're getting metadata straight from the source over TLS, it's
redundant.

With this commit, SAML Service can configure an x509 certificate for a
given source if double-verification is desired. Otherwise, it relies
solely on TLS for security.