Filter specific entity requests by tag
A request /mdq/aaf-edugain-export/entities will return every entity
we're exporting to eduGAIN. That's great! What's not so great is that
a request for /mdq/aaf-edugain-export/{sha256}xyzabc will return the
xyzabc entity regardless of whether we're exporting it to eduGAIN or
not. SAML service just ignores the prefix and returns any entity in it's
database.

This commit changes the behaviour so that it only returns an entity
under a specific tag if the entity actually has that tag.